Revert "Move flowSetupTimer from CommonMetrics to ExecMetrics (#2138)"
This reverts commit 3d13a95e46675377576d076c68d090aaaef90256.
