+++ azkaban-common/src/main/java/azkaban/metrics/CommonMetrics.java
@@ -19,6 +19,8 @@ package azkaban.metrics;
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
+import com.codahale.metrics.Timer;
+import java.util.concurrent.atomic.AtomicLong;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
@@ -39,6 +41,8 @@ public class CommonMetrics {
   public static final String SUBMIT_FLOW_SKIP_METER_NAME = "submit-flow-skip-meter";
   public static final String OOM_WAITING_JOB_COUNT_NAME = "OOM-waiting-job-count";
   public static final String QUEUE_WAIT_HISTOGRAM_NAME = "queue-wait-histogram";
+  public static final String FLOW_SETUP_TIMER_NAME = "flow-setup-timer";
+
 
   private Counter OOMWaitingJobCount;
   private final MetricsManager metricsManager;
@@ -51,6 +55,7 @@ public class CommonMetrics {
   private Meter submitFlowFailMeter;
   private Meter submitFlowSkipMeter;
   private Histogram queueWaitMeter;
+  private Timer flowSetupTimer;
 
   @Inject
   public CommonMetrics(final MetricsManager metricsManager) {
@@ -69,6 +74,7 @@ public class CommonMetrics {
     this.submitFlowSkipMeter = this.metricsManager.addMeter(SUBMIT_FLOW_SKIP_METER_NAME);
     this.OOMWaitingJobCount = this.metricsManager.addCounter(OOM_WAITING_JOB_COUNT_NAME);
     this.queueWaitMeter = this.metricsManager.addHistogram(QUEUE_WAIT_HISTOGRAM_NAME);
+    this.flowSetupTimer = this.metricsManager.addTimer(FLOW_SETUP_TIMER_NAME);
   }
 
   /**
@@ -147,7 +153,10 @@ public class CommonMetrics {
    *
    * @param time queue wait time for a flow.
    */
-  public void addQueueWait(final long time) {
-    this.queueWaitMeter.update(time);
-  }
+  public void addQueueWait(long time) { this.queueWaitMeter.update(time); }
+
+  /**
+   * @return the {@link Timer.Context} for the timer.
+   */
+  public Timer.Context getFlowSetupTimerContext() { return this.flowSetupTimer.time(); }
 }
